requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
httpx
python-dotenv
